dom/tests/mochitest/geolocation/test_mozsettings.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=478911
-->
<head>
  <title>Test for getCurrentPosition </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=777594">Mozilla Bug 777594</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
resume_geolocationProvider();
force_prompt(true);


SpecialPowers.setBoolPref("dom.mozSettings.enabled", true);
SpecialPowers.addPermission("settings-read", true, document);
SpecialPowers.addPermission("settings-write", true, document);
SpecialPowers.Cu.import("resource://gre/modules/SettingsChangeNotifier.jsm");

ok(navigator.geolocation, "get geolocation object");

toggleGeolocationSetting(false, function() {
    ok(true, "turned off geolocation via mozSettings");
    setTimeout(function() {
	navigator.geolocation.getCurrentPosition(successCallbackAfterMozsettingOff,
                                                 failureCallbackAfterMozsettingOff);
      }, 500); // need to wait a bit for all of these async callbacks to finish
  });

function successCallbackAfterMozsettingOff(position) {
  ok(false, "Success callback should not have been called after setting geolocation.enabled to false.");

  toggleGeolocationSetting(true, function() {
      ok(true, "turned on geolocation via mozSettings");
      setTimeout(function() {
         navigator.geolocation.getCurrentPosition(successCallbackAfterMozsettingOn,
                                                  failureCallbackAfterMozsettingOn);
        }, 500); // need to wait a bit for all of these async callbacks to finish
    });
}

function failureCallbackAfterMozsettingOff(error) {
  ok(true, "Geolocation didn't work after setting geolocation.enabled to false.");

  toggleGeolocationSetting(true, function() {
      ok(true, "turned on geolocation via mozSettings");
      setTimeout(function() {
         navigator.geolocation.getCurrentPosition(successCallbackAfterMozsettingOn,
                                                  failureCallbackAfterMozsettingOn);
        }, 500); // need to wait a bit for all of these async callbacks to finish
    });
}

function successCallbackAfterMozsettingOn(position) {
  ok(true, "Geolocation worked after setting geolocation.enabled to true.");

  reset_prompt();
  SimpleTest.finish();
}

function failureCallbackAfterMozsettingOn(error) {
  ok(false, "Geolocation didn't work after setting geolocation.enabled to true.");

  reset_prompt();
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
